fixed shutdown errors
It appears spring was not waiting for the scheduled tasks to finish running before closing the JPA entity manager. So we always ended up with an error being logged about the EntityManager already being closed.

Taking over control of the scheduled thread, we are blocking on shutdown so the scheduled task will now finish properly and then the JPA stuff is all good to be shutdown properly.

#funtimes